feat: added migration for tests
