+++ app/__init__.py
@@ -20,6 +20,8 @@ def create_app(config_name):
     app.config.from_object(config[config_name])
     config[config_name].init_app(app)
     bootstrap.init_app(app)
+    db.init_app(app)
+
     pagedown.init_app(app)
     CSRFProtect(app)
