+++ app/first_migration.py
@@ -1,0 +1,4 @@
+from flask_migrate import Migrate
+from app import create_app, db
+app = create_app('development' or 'default')
+migrate = Migrate(app, db)
