+++ business_card.py
@@ -1,9 +1,12 @@
 import os
-from app import create_app
+from app import create_app, db
+from flask_migrate import Migrate
+from app.models import admin, prospect
 
 app = create_app(os.getenv('FLASK_CONFIG') or 'default')
+migrate = Migrate(app, db)
 
 
 @app.shell_context_processor
 def make_shell_context():
-    return dict()
+    return dict(db=db, admin=admin, prospect=prospect)
